♻️ | refactor | A edição do livro não estava funcionando.
Diminuição de responsabilidade do componente cardsBooks, criando-se um componente aparte o 'editPopover.vue'.
