+++ client/dashboard-administracao/components.d.ts
@@ -22,6 +22,7 @@ declare module 'vue' {
     Dialog: typeof import('primevue/dialog')['default']
     Divider: typeof import('primevue/divider')['default']
     DynamicPhotoFrame: typeof import('./src/components/DynamicPhotoFrame.vue')['default']
+    EditPopover: typeof import('./src/components/bookView/editPopover.vue')['default']
     FloatLabel: typeof import('primevue/floatlabel')['default']
     HelloWorld: typeof import('./src/components/HelloWorld.vue')['default']
     IconCommunity: typeof import('./src/components/icons/IconCommunity.vue')['default']
